getting glsl import in sfe to work
