+++ packages/core/src/index.ts
@@ -37,3 +37,5 @@ export { SmartFilterOptimizer } from "./optimization/smartFilterOptimizer.js";
 export * from "./utils/textureLoaders.js";
 export * from "./serialization/index.js";
 export * from "./editorUtils/editableInPropertyPage.js";
+export * from "./utils/buildTools/shaderParsing.js";
+export * from "./utils/buildTools/convertShaderToSerializedBlockDefinition.js";
